@@ -9,37 +9,37 @@
     "prepublish": "npm run lint && npm run build"
   },
   "dependencies": {
-    "babel-polyfill": "^6.6.1",
-    "html-loader": "^0.4.3",
-    "json-loader": "^0.5.4",
-    "markdown-loader": "^0.1.7",
-    "raw-loader": "^0.5.1",
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
-    "react-redux": "^4.4.1",
-    "redux": "^3.3.1",
+    "babel-polyfill": "^6.16.0",
+    "html-loader": "^0.4.4",
+    "json-loader": "0.5.4",
+    "markdown-loader": "0.1.7",
+    "raw-loader": "0.5.1",
+    "react": "^15.4.1",
+    "react-dom": "^15.4.1",
+    "react-redux": "^4.4.6",
+    "redux": "^3.6.0",
     "redux-form": "file:../../",
-    "redux-form-website-template": "0.0.39"
+    "redux-form-website-template": "0.0.41"
   },
   "devDependencies": {
-    "babel-core": "^6.3.15",
-    "babel-eslint": "^5.0.0-beta4",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "cross-env": "^1.0.7",
-    "eslint": "^1.6.0",
-    "eslint-config-rackt": "^1.1.1",
-    "eslint-loader": "^1.3.0",
-    "eslint-plugin-babel": "^3.0.0",
-    "eslint-plugin-react": "^3.11.3",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "redbox-react": "^1.2.2",
-    "rimraf": "^2.4.3",
-    "webpack": "^1.12.9",
-    "webpack-dev-middleware": "^1.4.0",
-    "webpack-hot-middleware": "^2.6.0"
+    "babel-core": "^6.18.2",
+    "babel-eslint": "^7.1.1",
+    "babel-loader": "^6.2.8",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-react": "^6.16.0",
+    "cross-env": "^3.1.3",
+    "eslint": "^3.11.1",
+    "eslint-config-rackt": "1.1.1",
+    "eslint-loader": "^1.6.1",
+    "eslint-plugin-babel": "^4.0.0",
+    "eslint-plugin-react": "^6.7.1",
+    "eventsource-polyfill": "0.9.6",
+    "express": "^4.14.0",
+    "extract-text-webpack-plugin": "1.0.1",
+    "redbox-react": "^1.3.3",
+    "rimraf": "^2.5.4",
+    "webpack": "^1.13.3",
+    "webpack-dev-middleware": "^1.8.4",
+    "webpack-hot-middleware": "^2.13.2"
   }
 }
