@@ -2,7 +2,7 @@ import { Component, PropTypes, createElement } from 'react'
 import hoistStatics from 'hoist-non-react-statics'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
-import { mapValues } from 'lodash'
+import { mapValues, merge } from 'lodash'
 import isPromise from 'is-promise'
 import getDisplayName from './util/getDisplayName'
 import * as importedActions from './actions'
@@ -13,6 +13,7 @@ import asyncValidation from './asyncValidation'
 import defaultShouldAsyncValidate from './defaultShouldAsyncValidate'
 import defaultShouldValidate from './defaultShouldValidate'
 import plain from './structure/plain'
+import generateValidator from './generateValidator'
 import createIsValid from './selectors/isValid'
 
 const isClassComponent = Component => Boolean(
@@ -90,6 +91,7 @@ const createReduxForm =
         keepDirtyOnReinitialize: false,
         getFormState: state => getIn(state, 'form'),
         pure: true,
+        forceUnregisterOnUnmount: false,
         ...initialConfig
       }
 
@@ -106,6 +108,11 @@ const createReduxForm =
             this.register = this.register.bind(this)
             this.unregister = this.unregister.bind(this)
             this.submitCompleted = this.submitCompleted.bind(this)
+            this.submitFailed = this.submitFailed.bind(this)
+            this.fieldValidators = {}
+            this.lastFieldValidatorKeys = []
+            this.fieldWarners = {}
+            this.lastFieldWarnerKeys = []
 
             instances++
           }
@@ -119,7 +126,8 @@ const createReduxForm =
                 getValues: this.getValues,
                 sectionPrefix: undefined,
                 register: this.register,
-                unregister: this.unregister
+                unregister: this.unregister,
+                registerInnerOnSubmit: innerOnSubmit => this.innerOnSubmit = innerOnSubmit
               }
             }
           }
@@ -132,7 +140,7 @@ const createReduxForm =
                 this.props.initialize(nextProps.initialValues, keepDirty)
               }
             } else if (this.props.initialValues && (!this.props.initialized || enableReinitialize)) {
-              this.props.initialize(this.props.initialValues)
+              this.props.initialize(this.props.initialValues, this.props.keepDirtyOnReinitialize)
             }
           }
 
@@ -147,7 +155,7 @@ const createReduxForm =
 
           submitIfNeeded(nextProps) {
             const { clearSubmit, triggerSubmit } = this.props
-            if(!triggerSubmit && nextProps.triggerSubmit) {
+            if (!triggerSubmit && nextProps.triggerSubmit) {
               clearSubmit()
               this.submit()
             }
@@ -155,24 +163,29 @@ const createReduxForm =
 
           validateIfNeeded(nextProps) {
             const { shouldValidate, validate, values } = this.props
-            if (validate) {
+            const fieldLevelValidate = this.generateValidator()
+            if (validate || fieldLevelValidate) {
               const initialRender = nextProps === undefined
+              const fieldValidatorKeys = Object.keys(this.fieldValidators)
               const shouldValidateResult = shouldValidate({
                 values,
                 nextProps,
                 props: this.props,
                 initialRender,
+                lastFieldValidatorKeys: this.lastFieldValidatorKeys,
+                fieldValidatorKeys,
                 structure
               })
 
               if (shouldValidateResult) {
-                if (initialRender) {
-                  const { _error, ...nextSyncErrors } = validate(values, this.props)
-                  this.updateSyncErrorsIfNeeded(nextSyncErrors, _error)
-                } else {
-                  const { _error, ...nextSyncErrors } = validate(nextProps.values, nextProps)
-                  this.updateSyncErrorsIfNeeded(nextSyncErrors, _error)
-                }
+                const propsToValidate = initialRender ? this.props : nextProps
+                const { _error, ...nextSyncErrors } = merge(
+                  validate ? validate(propsToValidate.values, propsToValidate) || {} : {},
+                  fieldLevelValidate ?
+                  fieldLevelValidate(propsToValidate.values, propsToValidate) || {} : {}
+                )
+                this.lastFieldValidatorKeys = fieldValidatorKeys
+                this.updateSyncErrorsIfNeeded(nextSyncErrors, _error)
               }
             }
           }
@@ -187,17 +200,29 @@ const createReduxForm =
           }
 
           warnIfNeeded(nextProps) {
-            const { warn, values } = this.props
-            if (warn) {
-              if (nextProps) {
-                // not initial render
-                if (!deepEqual(values, nextProps.values)) {
-                  const { _warning, ...nextSyncWarnings } = warn(nextProps.values, nextProps)
-                  this.updateSyncWarningsIfNeeded(nextSyncWarnings, _warning)
-                }
-              } else {
-                // initial render
-                const { _warning, ...nextSyncWarnings } = warn(values, this.props)
+            const { shouldValidate, warn, values } = this.props
+            const fieldLevelWarn = this.generateWarner()
+            if (warn || fieldLevelWarn) {
+              const initialRender = nextProps === undefined
+              const fieldWarnerKeys = Object.keys(this.fieldWarners)
+              const shouldWarnResult = shouldValidate({
+                values,
+                nextProps,
+                props: this.props,
+                initialRender,
+                lastFieldValidatorKeys: this.lastFieldWarnerKeys,
+                fieldValidatorKeys: fieldWarnerKeys,
+                structure
+              })
+
+              if (shouldWarnResult) {
+                const propsToWarn = initialRender ? this.props : nextProps
+                const { _warning, ...nextSyncWarnings } = merge(
+                  warn ? warn(propsToWarn.values, propsToWarn) : {},
+                  fieldLevelWarn ?
+                    fieldLevelWarn(propsToWarn.values, propsToWarn) : {}
+                )
+                this.lastFieldWarnerKeys = fieldWarnerKeys
                 this.updateSyncWarningsIfNeeded(nextSyncWarnings, _warning)
               }
             }
@@ -251,18 +276,48 @@ const createReduxForm =
             return this.props.pristine
           }
 
-          register(name, type) {
+          register(name, type, getValidator, getWarner) {
             this.props.registerField(name, type)
+            if (getValidator) {
+              this.fieldValidators[ name ] = getValidator
+            }
+            if (getWarner) {
+              this.fieldWarners[ name ] = getWarner
+            }
           }
 
           unregister(name) {
-            if (!this.destroyed && (!this.unmounted || !instances)) {
+            if ((this.props.destroyOnUnmount || this.props.forceUnregisterOnUnmount) && !this.destroyed && (!this.unmounted || !instances)) {
               this.props.unregisterField(name)
+              delete this.fieldValidators[ name ]
+              delete this.fieldWarners[ name ]
             }
           }
 
           getFieldList() {
             return this.props.registeredFields.map((field) => getIn(field, 'name'))
+          }
+
+          generateValidator() {
+            const validators = {}
+            Object.keys(this.fieldValidators).forEach(name => {
+              const validator = this.fieldValidators[ name ]()
+              if (validator) {
+                validators[ name ] = validator
+              }
+            })
+            return Object.keys(validators).length ? generateValidator(validators, structure) : undefined
+          }
+
+          generateWarner() {
+            const warners = {}
+            Object.keys(this.fieldWarners).forEach(name => {
+              const warner = this.fieldWarners[ name ]()
+              if (warner) {
+                warners[ name ] = warner
+              }
+            })
+            return Object.keys(warners).length ? generateValidator(warners, structure) : undefined
           }
 
           asyncValidate(name, value) {
@@ -308,12 +363,17 @@ const createReduxForm =
             return result
           }
 
+          submitFailed(error) {
+            delete this.submitPromise
+            throw error
+          }
+
           listenToSubmit(promise) {
             if (!isPromise(promise)) {
               return promise
             }
             this.submitPromise = promise
-            return promise.then(this.submitCompleted)
+            return promise.then(this.submitCompleted, this.submitFailed)
           }
 
           submit(submitOrEvent) {
@@ -322,15 +382,21 @@ const createReduxForm =
             if (!submitOrEvent || silenceEvent(submitOrEvent)) {
               // submitOrEvent is an event: fire submit if not already submitting
               if (!this.submitPromise) {
-                return this.listenToSubmit(handleSubmit(checkSubmit(onSubmit),
-                  this.props, this.isValid(), this.asyncValidate, this.getFieldList()))
+                if (this.innerOnSubmit) {
+                  // will call "submitOrEvent is the submit function" block below
+                  return this.innerOnSubmit()
+                } else {
+                  return this.listenToSubmit(handleSubmit(checkSubmit(onSubmit),
+                    this.props, this.props.validExceptSubmit, this.asyncValidate, this.getFieldList()))
+                }
               }
             } else {
               // submitOrEvent is the submit function: return deferred submit thunk
-              return silenceEvents(() =>
-              !this.submitPromise &&
-              this.listenToSubmit(handleSubmit(checkSubmit(submitOrEvent),
-                this.props, this.isValid(), this.asyncValidate, this.getFieldList())))
+              return silenceEvents(() => {
+                return !this.submitPromise &&
+                  this.listenToSubmit(handleSubmit(checkSubmit(submitOrEvent),
+                    this.props, this.props.validExceptSubmit, this.asyncValidate, this.getFieldList()))
+              })
             }
           }
 
@@ -360,6 +426,7 @@ const createReduxForm =
               change,
               destroy,
               destroyOnUnmount,
+              forceUnregisterOnUnmount,
               dirty,
               dispatch,
               enableReinitialize,
@@ -399,6 +466,7 @@ const createReduxForm =
               updateSyncErrors,
               updateSyncWarnings,
               valid,
+              validExceptSubmit,
               values,
               warning,
               ...rest
@@ -446,6 +514,7 @@ const createReduxForm =
         }
         Form.propTypes = {
           destroyOnUnmount: PropTypes.bool,
+          forceUnregisterOnUnmount: PropTypes.bool,
           form: PropTypes.string.isRequired,
           initialValues: PropTypes.object,
           getFormState: PropTypes.func,
@@ -466,14 +535,15 @@ const createReduxForm =
             const { form, getFormState, initialValues, enableReinitialize, keepDirtyOnReinitialize } = props
             const formState = getIn(getFormState(state) || empty, form) || empty
             const stateInitial = getIn(formState, 'initial')
+            const initialized = !!stateInitial
 
-            const shouldUpdateInitialValues = enableReinitialize && !deepEqual(initialValues, stateInitial)
+            const shouldUpdateInitialValues = enableReinitialize && initialized && !deepEqual(initialValues, stateInitial)
             const shouldResetValues = shouldUpdateInitialValues && !keepDirtyOnReinitialize
 
             let initial = initialValues || stateInitial || empty
 
             if (shouldUpdateInitialValues) {
-              initial = stateInitial
+              initial = stateInitial || empty
             }
 
             let values = getIn(formState, 'values') || initial
@@ -487,7 +557,8 @@ const createReduxForm =
             const syncErrors = getIn(formState, 'syncErrors') || {}
             const syncWarnings = getIn(formState, 'syncWarnings') || {}
             const registeredFields = getIn(formState, 'registeredFields') || []
-            const valid = isValid(form, getFormState)(state)
+            const valid = isValid(form, getFormState, false)(state)
+            const validExceptSubmit = isValid(form, getFormState, true)(state)
             const anyTouched = !!getIn(formState, 'anyTouched')
             const submitting = !!getIn(formState, 'submitting')
             const submitFailed = !!getIn(formState, 'submitFailed')
@@ -501,7 +572,7 @@ const createReduxForm =
               asyncValidating: getIn(formState, 'asyncValidating') || false,
               dirty: !pristine,
               error,
-              initialized: !!stateInitial,
+              initialized,
               invalid: !valid,
               pristine,
               registeredFields,
@@ -513,6 +584,7 @@ const createReduxForm =
               triggerSubmit,
               values,
               valid,
+              validExceptSubmit,
               warning
             }
           },
