@@ -35,9 +35,9 @@ let render = () => {
          * This <App/> component only provides the site wrapper.
          * Remove it on your dev server if you wish. It will not affect the functionality.
          */
-        version="6.2.0"
+        version="6.4.3"
         path="/examples/initializeFromState"
-        breadcrumbs={generateExampleBreadcrumbs('initializeFromState', 'Initialize From State Example', '6.2.0')}>
+        breadcrumbs={generateExampleBreadcrumbs('initializeFromState', 'Initialize From State Example', '6.4.3')}>
 
         <Markdown content={readme}/>
 
