@@ -3,23 +3,46 @@ import { Field, reduxForm } from 'redux-form'
 import DropdownList from 'react-widgets/lib/DropdownList'
 import SelectList from 'react-widgets/lib/SelectList'
 import Multiselect from 'react-widgets/lib/Multiselect'
+import DateTimePicker from 'react-widgets/lib/DateTimePicker'
+import moment from 'moment'
+import momentLocaliser from 'react-widgets/lib/localizers/moment'
+
 import 'react-widgets/dist/css/react-widgets.css'
+
+momentLocaliser(moment)
 
 const colors = [ { color: 'Red', value: 'ff0000' },
   { color: 'Green', value: '00ff00' },
   { color: 'Blue', value: '0000ff' } ]
 
-const renderDropdownList = ({ input, ...rest }) =>
-  <DropdownList {...input} {...rest}/>
+const renderDropdownList = ({ input, data, valueField, textField }) =>
+  <DropdownList {...input}
+    data={data}
+    valueField={valueField}
+    textField={textField}
+    onChange={input.onChange} />
 
-const renderMultiselect = ({ input, ...rest }) =>
+const renderMultiselect = ({ input, data, valueField, textField }) =>
   <Multiselect {...input}
     onBlur={() => input.onBlur()}
     value={input.value || []} // requires value to be an array
-    {...rest}/>
+    data={data}
+    valueField={valueField}
+    textField={textField}
+  />
 
-const renderSelectList = ({ input, ...rest }) =>
-  <SelectList {...input} onBlur={() => input.onBlur()} {...rest}/>
+const renderSelectList = ({ input, data }) =>
+  <SelectList {...input}
+    onBlur={() => input.onBlur()}
+    data={data} />
+
+const renderDateTimePicker = ({ input: { onChange, value }, showTime }) =>
+  <DateTimePicker
+    onChange={onChange}
+    format="DD MMM YYYY"
+    time={showTime}
+    value={value && new Date(value) || null}
+  />
 
 let ReactWidgetsForm = props => {
   const { handleSubmit, pristine, reset, submitting } = props
@@ -47,6 +70,14 @@ let ReactWidgetsForm = props => {
           name="sex"
           component={renderSelectList}
           data={[ 'male', 'female' ]}/>
+      </div>
+      <div>
+        <label>DOB</label>
+        <Field
+          name="dob"
+          showTime={false}
+          component={renderDateTimePicker}
+        />
       </div>
       <div>
         <button type="submit" disabled={pristine || submitting}>Submit</button>
