@@ -2,12 +2,14 @@ import { Component, PropTypes, createElement } from 'react'
 import { connect } from 'react-redux'
 import createFieldProps from './createFieldProps'
 import plain from './structure/plain'
+import onChangeValue from './events/onChangeValue'
+import { dataKey } from './util/eventConsts'
 
 const propsToNotUpdateFor = [
   '_reduxForm'
 ]
 
-const createConnectedField = ({ deepEqual, getIn }) => {
+const createConnectedField = ({ deepEqual, getIn, toJS }) => {
 
   const getSyncError = (syncErrors, name) => {
     const error = plain.getIn(syncErrors, name)
@@ -24,6 +26,16 @@ const createConnectedField = ({ deepEqual, getIn }) => {
   }
 
   class ConnectedField extends Component {
+    constructor(props) {
+      super(props)
+
+      this.handleChange = this.handleChange.bind(this)
+      this.handleFocus = this.handleFocus.bind(this)
+      this.handleBlur = this.handleBlur.bind(this)
+      this.handleDragStart = this.handleDragStart.bind(this)
+      this.handleDrop = this.handleDrop.bind(this)
+    }
+
     shouldComponentUpdate(nextProps) {
       const nextPropsKeys = Object.keys(nextProps)
       const thisPropsKeys = Object.keys(this.props)
@@ -44,19 +56,127 @@ const createConnectedField = ({ deepEqual, getIn }) => {
       return this.refs.renderedComponent
     }
 
+    handleChange(event) {
+      const { name, dispatch, parse, normalize, onChange, _reduxForm, value: previousValue } = this.props
+      const newValue = onChangeValue(event, { name, parse, normalize })
+
+      let defaultPrevented = false
+      if (onChange) {
+        onChange({
+          ...event,
+          preventDefault: () => {
+            defaultPrevented = true
+            return event.preventDefault()
+          }
+        }, newValue, previousValue)
+      }
+      if (!defaultPrevented) {
+        // dispatch change action
+        dispatch(_reduxForm.change(name, newValue))
+      }
+    }
+
+    handleFocus(event) {
+      const { name, dispatch, onFocus, _reduxForm } = this.props
+
+      let defaultPrevented = false
+      if (onFocus) {
+        onFocus({
+          ...event,
+          preventDefault: () => {
+            defaultPrevented = true
+            return event.preventDefault()
+          }
+        })
+      }
+
+      if (!defaultPrevented) {
+        dispatch(_reduxForm.focus(name))
+      }
+    }
+
+    handleBlur(event) {
+      const { name, dispatch, parse, normalize, onBlur, _reduxForm, value: previousValue } = this.props
+      const newValue = onChangeValue(event, { name, parse, normalize })
+
+      let defaultPrevented = false
+      if (onBlur) {
+        onBlur({
+          ...event,
+          preventDefault: () => {
+            defaultPrevented = true
+            return event.preventDefault()
+          }
+        }, newValue, previousValue)
+      }
+
+      if (!defaultPrevented) {
+        // dispatch blur action
+        dispatch(_reduxForm.blur(name, newValue))
+
+        // call post-blur callback
+        if (_reduxForm.asyncValidate) {
+          _reduxForm.asyncValidate(name, newValue)
+        }
+      }
+    }
+
+    handleDragStart(event) {
+      const { onDragStart, value } = this.props
+      event.dataTransfer.setData(dataKey, value == null ? '' : value)
+
+      if (onDragStart) {
+        onDragStart(event)
+      }
+    }
+
+    handleDrop(event) {
+      const { name, dispatch, onDrop, _reduxForm, value: previousValue } = this.props
+      const newValue = event.dataTransfer.getData(dataKey)
+
+      let defaultPrevented = false
+      if (onDrop) {
+        onDrop({
+          ...event,
+          preventDefault: () => {
+            defaultPrevented = true
+            return event.preventDefault()
+          }
+        }, newValue, previousValue)
+      }
+
+      if (!defaultPrevented) {
+        // dispatch change action
+        dispatch(_reduxForm.change(name, newValue))
+        event.preventDefault()
+      }
+    }
+
     render() {
-      const { component, withRef, name, _reduxForm, ...rest } = this.props
-      const { asyncValidate, blur, change, focus } = _reduxForm
-      const { custom, ...props } = createFieldProps(getIn,
+      const {
+        component,
+        withRef,
+        name,
+        // remove props that are part of redux internals:
+        _reduxForm,  // eslint-disable-line no-unused-vars
+        normalize,   // eslint-disable-line no-unused-vars
+        onBlur,      // eslint-disable-line no-unused-vars
+        onChange,    // eslint-disable-line no-unused-vars
+        onFocus,     // eslint-disable-line no-unused-vars
+        onDragStart, // eslint-disable-line no-unused-vars
+        onDrop,      // eslint-disable-line no-unused-vars
+        ...rest
+      } = this.props
+      const { custom, ...props } = createFieldProps({ getIn, toJS },
         name,
         {
           ...rest,
-          name,
-          blur,
-          change,
-          focus
-        },
-        asyncValidate
+          onBlur: this.handleBlur,
+          onChange: this.handleChange,
+          onDrop: this.handleDrop,
+          onDragStart: this.handleDragStart,
+          onFocus: this.handleFocus
+        }
       )
       if (withRef) {
         custom.ref = 'renderedComponent'
@@ -94,6 +214,7 @@ const createConnectedField = ({ deepEqual, getIn }) => {
         pristine,
         state: getIn(formState, `fields.${name}`),
         submitError: getIn(formState, `submitErrors.${name}`),
+        submitFailed: getIn(formState, 'submitFailed'),
         submitting,
         syncError,
         syncWarning,
