@@ -7,86 +7,73 @@ import immutableExpectations from '../structure/immutable/expectations'
 import addExpectations from './addExpectations'
 
 const describeCreateFieldProps = (name, structure, expect) => {
-  const { empty, getIn, fromJS } = structure
+  const { empty, getIn, fromJS, toJS } = structure
 
   describe(name, () => {
     it('should pass value through', () => {
-      expect(createFieldProps(getIn, 'foo', { value: 'hello' }).input.value).toBe('hello')
+      expect(createFieldProps({ getIn, toJS }, 'foo', { value: 'hello' }).input.value).toBe('hello')
     })
 
     it('should pass dirty/pristine through', () => {
-      expect(createFieldProps(getIn, 'foo', { dirty: false, pristine: true }).meta.dirty).toBe(false)
-      expect(createFieldProps(getIn, 'foo', { dirty: false, pristine: true }).meta.pristine).toBe(true)
-      expect(createFieldProps(getIn, 'foo', { dirty: true, pristine: false }).meta.dirty).toBe(true)
-      expect(createFieldProps(getIn, 'foo', { dirty: true, pristine: false }).meta.pristine).toBe(false)
+      expect(createFieldProps({ getIn, toJS }, 'foo', { dirty: false, pristine: true }).meta.dirty).toBe(false)
+      expect(createFieldProps({ getIn, toJS }, 'foo', { dirty: false, pristine: true }).meta.pristine).toBe(true)
+      expect(createFieldProps({ getIn, toJS }, 'foo', { dirty: true, pristine: false }).meta.dirty).toBe(true)
+      expect(createFieldProps({ getIn, toJS }, 'foo', { dirty: true, pristine: false }).meta.pristine).toBe(false)
     })
 
     it('should provide onBlur', () => {
-      const blur = createSpy()
-      const dispatch = createSpy()
-      const normalize = createSpy((name, value) => value).andCallThrough()
-      expect(blur).toNotHaveBeenCalled()
-      const result = createFieldProps(getIn, 'foo', { value: 'bar', blur, dispatch, normalize })
+      const onBlur = createSpy()
+      expect(onBlur).toNotHaveBeenCalled()
+      const result = createFieldProps({ getIn, toJS }, 'foo', { value: 'bar', onBlur })
       expect(result.input.onBlur).toBeA('function')
-      expect(blur).toNotHaveBeenCalled()
-      expect(dispatch).toNotHaveBeenCalled()
+      expect(onBlur).toNotHaveBeenCalled()
       result.input.onBlur('rabbit')
-      expect(normalize).toHaveBeenCalled()
-      expect(blur)
+      expect(onBlur)
         .toHaveBeenCalled()
-        .toHaveBeenCalledWith('foo', 'rabbit')
-      expect(dispatch).toHaveBeenCalled()
+        .toHaveBeenCalledWith('rabbit')
     })
 
     it('should provide onChange', () => {
-      const change = createSpy()
-      const dispatch = createSpy()
-      const normalize = createSpy((name, value) => value).andCallThrough()
-      expect(change).toNotHaveBeenCalled()
-      const result = createFieldProps(getIn, 'foo', { value: 'bar', change, dispatch, normalize })
+      const onChange = createSpy()
+      expect(onChange).toNotHaveBeenCalled()
+      const result = createFieldProps({ getIn, toJS }, 'foo', { value: 'bar', onChange })
       expect(result.input.onChange).toBeA('function')
-      expect(change).toNotHaveBeenCalled()
-      expect(dispatch).toNotHaveBeenCalled()
+      expect(onChange).toNotHaveBeenCalled()
       result.input.onChange('rabbit')
-      expect(normalize).toHaveBeenCalled()
-      expect(change)
+      expect(onChange)
         .toHaveBeenCalled()
-        .toHaveBeenCalledWith('foo', 'rabbit')
-      expect(dispatch).toHaveBeenCalled()
+        .toHaveBeenCalledWith('rabbit')
     })
 
     it('should provide onFocus', () => {
-      const focus = createSpy()
-      const dispatch = createSpy()
-      expect(focus).toNotHaveBeenCalled()
-      const result = createFieldProps(getIn, 'foo', { value: 'bar', dispatch, focus })
+      const onFocus = createSpy()
+      expect(onFocus).toNotHaveBeenCalled()
+      const result = createFieldProps({ getIn, toJS }, 'foo', { value: 'bar', onFocus })
       expect(result.input.onFocus).toBeA('function')
-      expect(focus).toNotHaveBeenCalled()
-      expect(dispatch).toNotHaveBeenCalled()
+      expect(onFocus).toNotHaveBeenCalled()
       result.input.onFocus('rabbit')
-      expect(focus)
-        .toHaveBeenCalled()
-        .toHaveBeenCalledWith('foo')
-      expect(dispatch).toHaveBeenCalled()
+      expect(onFocus).toHaveBeenCalled()
     })
 
     it('should provide onDragStart', () => {
-      const result = createFieldProps(getIn, 'foo', { value: 'bar' })
+      const onDragStart = createSpy()
+      const result = createFieldProps({ getIn, toJS }, 'foo', { value: 'bar', onDragStart })
       expect(result.input.onDragStart).toBeA('function')
     })
 
     it('should provide onDrop', () => {
-      const result = createFieldProps(getIn, 'foo', { value: 'bar' })
+      const onDrop = createSpy()
+      const result = createFieldProps({ getIn, toJS }, 'foo', { value: 'bar', onDrop })
       expect(result.input.onDrop).toBeA('function')
     })
 
     it('should read active from state', () => {
-      const inactiveResult = createFieldProps(getIn, 'foo', {
+      const inactiveResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty
       })
       expect(inactiveResult.meta.active).toBe(false)
-      const activeResult = createFieldProps(getIn, 'foo', {
+      const activeResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: fromJS({
           active: true
@@ -96,53 +83,65 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should pass along submitting flag', () => {
-      const notSubmittingResult = createFieldProps(getIn, 'foo', {
+      const notSubmittingResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar'
       })
       expect(notSubmittingResult.meta.submitting).toBe(false)
-      const submittingResult = createFieldProps(getIn, 'foo', {
+      const submittingResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         submitting: true
       })
       expect(submittingResult.meta.submitting).toBe(true)
     })
 
+    it('should pass along submitFailed flag', () => {
+      const notFailedResult = createFieldProps({ getIn, toJS }, 'foo', {
+        value: 'bar'
+      })
+      expect(notFailedResult.meta.submitFailed).toBe(false)
+      const failedResult = createFieldProps({ getIn, toJS }, 'foo', {
+        value: 'bar',
+        submitFailed: true
+      })
+      expect(failedResult.meta.submitFailed).toBe(true)
+    })
+
     it('should pass along all custom state props', () => {
-      const pristineResult = createFieldProps(getIn, 'foo', {
+      const pristineResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar'
       })
       expect(pristineResult.meta.customProp).toBe(undefined)
-      const customResult = createFieldProps(getIn, 'foo', {
+      const customResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
-        state: {
+        state: fromJS({
           customProp: 'my-custom-prop'
-        }
+        })
       })
       expect(customResult.meta.customProp).toBe('my-custom-prop')
     })
 
     it('should not override canonical props with custom props', () => {
-      const pristineResult = createFieldProps(getIn, 'foo', {
+      const pristineResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar'
       })
       expect(pristineResult.meta.customProp).toBe(undefined)
-      const customResult = createFieldProps(getIn, 'foo', {
+      const customResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         submitting: true,
-        state: {
+        state: fromJS({
           submitting: false
-        }
+        })
       })
       expect(customResult.meta.submitting).toBe(true)
     })
 
     it('should read touched from state', () => {
-      const untouchedResult = createFieldProps(getIn, 'foo', {
+      const untouchedResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty
       })
       expect(untouchedResult.meta.touched).toBe(false)
-      const touchedResult = createFieldProps(getIn, 'foo', {
+      const touchedResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: fromJS({
           touched: true
@@ -152,12 +151,12 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should read visited from state', () => {
-      const notVisitedResult = createFieldProps(getIn, 'foo', {
+      const notVisitedResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty
       })
       expect(notVisitedResult.meta.visited).toBe(false)
-      const visitedResult = createFieldProps(getIn, 'foo', {
+      const visitedResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: fromJS({
           visited: true
@@ -167,14 +166,14 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should read sync errors from prop', () => {
-      const noErrorResult = createFieldProps(getIn, 'foo', {
+      const noErrorResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty
       })
       expect(noErrorResult.meta.error).toNotExist()
       expect(noErrorResult.meta.valid).toBe(true)
       expect(noErrorResult.meta.invalid).toBe(false)
-      const errorResult = createFieldProps(getIn, 'foo', {
+      const errorResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty,
         syncError: 'This is an error'
@@ -185,12 +184,12 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should read sync warnings from prop', () => {
-      const noWarningResult = createFieldProps(getIn, 'foo', {
+      const noWarningResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty
       })
       expect(noWarningResult.meta.warning).toNotExist()
-      const warningResult = createFieldProps(getIn, 'foo', {
+      const warningResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty,
         syncWarning: 'This is an warning'
@@ -199,14 +198,14 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should read async errors from state', () => {
-      const noErrorResult = createFieldProps(getIn, 'foo', {
+      const noErrorResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty
       })
       expect(noErrorResult.meta.error).toNotExist()
       expect(noErrorResult.meta.valid).toBe(true)
       expect(noErrorResult.meta.invalid).toBe(false)
-      const errorResult = createFieldProps(getIn, 'foo', {
+      const errorResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty,
         syncError: 'This is an error'
@@ -217,14 +216,14 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should read submit errors from state', () => {
-      const noErrorResult = createFieldProps(getIn, 'foo', {
+      const noErrorResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty
       })
       expect(noErrorResult.meta.error).toNotExist()
       expect(noErrorResult.meta.valid).toBe(true)
       expect(noErrorResult.meta.invalid).toBe(false)
-      const errorResult = createFieldProps(getIn, 'foo', {
+      const errorResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty,
         submitError: 'This is an error'
@@ -235,14 +234,14 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should prioritize sync errors over async or submit errors', () => {
-      const noErrorResult = createFieldProps(getIn, 'foo', {
+      const noErrorResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty
       })
       expect(noErrorResult.meta.error).toNotExist()
       expect(noErrorResult.meta.valid).toBe(true)
       expect(noErrorResult.meta.invalid).toBe(false)
-      const errorResult = createFieldProps(getIn, 'foo', {
+      const errorResult = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         asyncError: 'async error',
         submitError: 'submit error',
@@ -254,7 +253,7 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should pass through other props', () => {
-      const result = createFieldProps(getIn, 'foo', {
+      const result = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty,
         someOtherProp: 'dog',
@@ -267,7 +266,7 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should pass through other props using props prop', () => {
-      const result = createFieldProps(getIn, 'foo', {
+      const result = createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty,
         props: {
@@ -282,16 +281,16 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should set checked for checkboxes', () => {
-      expect(createFieldProps(getIn, 'foo', {
+      expect(createFieldProps({ getIn, toJS }, 'foo', {
         state: empty,
         type: 'checkbox'
       }).input.checked).toBe(false)
-      expect(createFieldProps(getIn, 'foo', {
+      expect(createFieldProps({ getIn, toJS }, 'foo', {
         value: true,
         state: empty,
         type: 'checkbox'
       }).input.checked).toBe(true)
-      expect(createFieldProps(getIn, 'foo', {
+      expect(createFieldProps({ getIn, toJS }, 'foo', {
         value: false,
         state: empty,
         type: 'checkbox'
@@ -299,18 +298,18 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should set checked for radio buttons', () => {
-      expect(createFieldProps(getIn, 'foo', {
+      expect(createFieldProps({ getIn, toJS }, 'foo', {
         state: empty,
         type: 'radio',
         _value: 'bar'
       }).input.checked).toBe(false)
-      expect(createFieldProps(getIn, 'foo', {
+      expect(createFieldProps({ getIn, toJS }, 'foo', {
         value: 'bar',
         state: empty,
         type: 'radio',
         _value: 'bar'
       }).input.checked).toBe(true)
-      expect(createFieldProps(getIn, 'foo', {
+      expect(createFieldProps({ getIn, toJS }, 'foo', {
         value: 'baz',
         state: empty,
         type: 'radio',
@@ -319,7 +318,7 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should default value to [] for multi-selects', () => {
-      expect(createFieldProps(getIn, 'foo', {
+      expect(createFieldProps({ getIn, toJS }, 'foo', {
         state: empty,
         type: 'select-multiple'
       }).input.value)
@@ -328,7 +327,7 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should default value to undefined for file inputs', () => {
-      expect(createFieldProps(getIn, 'foo', {
+      expect(createFieldProps({ getIn, toJS }, 'foo', {
         state: empty,
         type: 'file'
       }).input.value)
@@ -336,10 +335,19 @@ const describeCreateFieldProps = (name, structure, expect) => {
     })
 
     it('should replace undefined value with empty string', () => {
-      const result = createFieldProps(getIn, 'foo', {
+      const result = createFieldProps({ getIn, toJS }, 'foo', {
         state: empty
       })
       expect(result.input.value).toBe('')
+    })
+
+    it('should not format value when format prop is null', () => {
+      const result = createFieldProps({ getIn, toJS }, 'foo', {
+        state: empty,
+        value: null,
+        format: null
+      })
+      expect(result.input.value).toBe(null)
     })
   })
 }
