@@ -152,7 +152,9 @@ describe('handleSubmit', () => {
         expect(stopSubmit).toNotHaveBeenCalled()
         expect(onSubmitFail)
           .toHaveBeenCalled()
-          .toHaveBeenCalledWith(values, dispatch)
+        expect(onSubmitFail.calls[0].arguments[0]).toEqual(values)
+        expect(onSubmitFail.calls[0].arguments[1]).toEqual(dispatch)
+        expect(onSubmitFail.calls[0].arguments[2]).toBe(null)
         expect(touch)
           .toHaveBeenCalled()
           .toHaveBeenCalledWith('foo', 'baz')
