@@ -1,31 +1,32 @@
-export const ARRAY_INSERT = 'redux-form/ARRAY_INSERT'
-export const ARRAY_MOVE = 'redux-form/ARRAY_MOVE'
-export const ARRAY_POP = 'redux-form/ARRAY_POP'
-export const ARRAY_PUSH = 'redux-form/ARRAY_PUSH'
-export const ARRAY_REMOVE = 'redux-form/ARRAY_REMOVE'
-export const ARRAY_REMOVE_ALL = 'redux-form/ARRAY_REMOVE_ALL'
-export const ARRAY_SHIFT = 'redux-form/ARRAY_SHIFT'
-export const ARRAY_SPLICE = 'redux-form/ARRAY_SPLICE'
-export const ARRAY_UNSHIFT = 'redux-form/ARRAY_UNSHIFT'
-export const ARRAY_SWAP = 'redux-form/ARRAY_SWAP'
-export const AUTOFILL = 'redux-form/AUTOFILL'
-export const BLUR = 'redux-form/BLUR'
-export const CHANGE = 'redux-form/CHANGE'
-export const CLEAR_SUBMIT = 'redux-form/CLEAR_SUBMIT'
-export const DESTROY = 'redux-form/DESTROY'
-export const FOCUS = 'redux-form/FOCUS'
-export const INITIALIZE = 'redux-form/INITIALIZE'
-export const REGISTER_FIELD = 'redux-form/REGISTER_FIELD'
-export const RESET = 'redux-form/RESET'
-export const SET_SUBMIT_FAILED = 'redux-form/SET_SUBMIT_FAILED'
-export const SET_SUBMIT_SUCCEEDED = 'redux-form/SET_SUBMIT_SUCCEEDED'
-export const START_ASYNC_VALIDATION = 'redux-form/START_ASYNC_VALIDATION'
-export const START_SUBMIT = 'redux-form/START_SUBMIT'
-export const STOP_ASYNC_VALIDATION = 'redux-form/STOP_ASYNC_VALIDATION'
-export const STOP_SUBMIT = 'redux-form/STOP_SUBMIT'
-export const SUBMIT = 'redux-form/SUBMIT'
-export const TOUCH = 'redux-form/TOUCH'
-export const UNREGISTER_FIELD = 'redux-form/UNREGISTER_FIELD'
-export const UNTOUCH = 'redux-form/UNTOUCH'
-export const UPDATE_SYNC_ERRORS = 'redux-form/UPDATE_SYNC_ERRORS'
-export const UPDATE_SYNC_WARNINGS = 'redux-form/UPDATE_SYNC_WARNINGS'
+export const ARRAY_INSERT = '@@redux-form/ARRAY_INSERT'
+export const ARRAY_MOVE = '@@redux-form/ARRAY_MOVE'
+export const ARRAY_POP = '@@redux-form/ARRAY_POP'
+export const ARRAY_PUSH = '@@redux-form/ARRAY_PUSH'
+export const ARRAY_REMOVE = '@@redux-form/ARRAY_REMOVE'
+export const ARRAY_REMOVE_ALL = '@@redux-form/ARRAY_REMOVE_ALL'
+export const ARRAY_SHIFT = '@@redux-form/ARRAY_SHIFT'
+export const ARRAY_SPLICE = '@@redux-form/ARRAY_SPLICE'
+export const ARRAY_UNSHIFT = '@@redux-form/ARRAY_UNSHIFT'
+export const ARRAY_SWAP = '@@redux-form/ARRAY_SWAP'
+export const AUTOFILL = '@@redux-form/AUTOFILL'
+export const BLUR = '@@redux-form/BLUR'
+export const CHANGE = '@@redux-form/CHANGE'
+export const CLEAR_SUBMIT = '@@redux-form/CLEAR_SUBMIT'
+export const CLEAR_ASYNC_ERROR = '@redux-form/CLEAR_ASYNC_ERROR'
+export const DESTROY = '@@redux-form/DESTROY'
+export const FOCUS = '@@redux-form/FOCUS'
+export const INITIALIZE = '@@redux-form/INITIALIZE'
+export const REGISTER_FIELD = '@@redux-form/REGISTER_FIELD'
+export const RESET = '@@redux-form/RESET'
+export const SET_SUBMIT_FAILED = '@@redux-form/SET_SUBMIT_FAILED'
+export const SET_SUBMIT_SUCCEEDED = '@@redux-form/SET_SUBMIT_SUCCEEDED'
+export const START_ASYNC_VALIDATION = '@@redux-form/START_ASYNC_VALIDATION'
+export const START_SUBMIT = '@@redux-form/START_SUBMIT'
+export const STOP_ASYNC_VALIDATION = '@@redux-form/STOP_ASYNC_VALIDATION'
+export const STOP_SUBMIT = '@@redux-form/STOP_SUBMIT'
+export const SUBMIT = '@@redux-form/SUBMIT'
+export const TOUCH = '@@redux-form/TOUCH'
+export const UNREGISTER_FIELD = '@@redux-form/UNREGISTER_FIELD'
+export const UNTOUCH = '@@redux-form/UNTOUCH'
+export const UPDATE_SYNC_ERRORS = '@@redux-form/UPDATE_SYNC_ERRORS'
+export const UPDATE_SYNC_WARNINGS = '@@redux-form/UPDATE_SYNC_WARNINGS'
