@@ -7,11 +7,16 @@ import createFormValueSelector from './formValueSelector'
 import createValues from './values'
 import createGetFormValues from './selectors/getFormValues'
 import createGetFormSyncErrors from './selectors/getFormSyncErrors'
+import createGetFormAsyncErrors from './selectors/getFormAsyncErrors'
 import createGetFormSubmitErrors from './selectors/getFormSubmitErrors'
 import createIsDirty from './selectors/isDirty'
 import createIsInvalid from './selectors/isInvalid'
 import createIsPristine from './selectors/isPristine'
 import createIsValid from './selectors/isValid'
+import createIsSubmitting from './selectors/isSubmitting'
+import createHasSubmitSucceeded from './selectors/hasSubmitSucceeded'
+import createHasSubmitFailed from './selectors/hasSubmitFailed'
+import Form from './Form'
 import FormSection from './FormSection'
 import SubmissionError from './SubmissionError'
 import propTypes from './propTypes'
@@ -25,15 +30,20 @@ const createAll = structure => ({
   Field: createField(structure),
   Fields: createFields(structure),
   FieldArray: createFieldArray(structure),
-  FormSection: FormSection,
+  Form,
+  FormSection,
   formValueSelector: createFormValueSelector(structure),
   getFormValues: createGetFormValues(structure),
   getFormSyncErrors: createGetFormSyncErrors(structure),
+  getFormAsyncErrors: createGetFormAsyncErrors(structure),
   getFormSubmitErrors: createGetFormSubmitErrors(structure),
   isDirty: createIsDirty(structure),
   isInvalid: createIsInvalid(structure),
   isPristine: createIsPristine(structure),
   isValid: createIsValid(structure),
+  isSubmitting: createIsSubmitting(structure),
+  hasSubmitSucceeded: createHasSubmitSucceeded(structure),
+  hasSubmitFailed: createHasSubmitFailed(structure),
   propTypes,
   reduxForm: createReduxForm(structure),
   reducer: createReducer(structure),
