@@ -17,7 +17,8 @@ const structure = {
     Iterable.isIndexed(value) ? value.toList() : value.toMap()),
   size: list => list ? list.size : 0,
   some: (iterable, callback) => Iterable.isIterable(iterable) ? iterable.some(callback) : false,
-  splice
+  splice,
+  toJS: value => Iterable.isIterable(value) ? value.toJS() : value
 }
 
 export default structure
