@@ -1,7 +1,7 @@
 import { isEqualWith } from 'lodash'
 
 const customizer = (obj, other) => {
-  if (obj == other) return true
+  if (obj === other) return true
   if ((obj == null || obj === '' || obj === false) &&
     (other == null || other === '' || other === false)) return true
 
